# fastingmate-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel API Documentation",
        "description": "Dokumentasi API untuk autentikasi pengguna",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": ""
        }
    ],
    "paths": {
...
        "/api/notifications": {
            "get": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Menampilkan daftar notifikasi pengguna",
                "operationId": "911565ad9aff9ba7eded3f189509e578",
                "responses": {
                    "200": {
                        "description": "Daftar notifikasi berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "user_id": {
                                                "type": "integer",
                                                "example": 5
                                            },
                                            "title": {
                                                "type": "string",
                                                "example": "Pengingat Puasa"
                                            },
                                            "message": {
                                                "type": "string",
                                                "example": "Jangan lupa sahur pukul 04:30"
                                            },
                                            "type": {
                                                "type": "string",
                                                "example": "reminder"
                                            },
                                            "is_read": {
                                                "type": "boolean",
                                                "example": false
                                            },
                                            "sent_at": {
                                                "type": "string",
                                                "format": "datetime",
                                                "example": "2025-03-21 08:00:00"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Membuat notifikasi baru",
                "operationId": "91f8e779a9e46cb5cdda969e5a928a94",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "message",
                                    "type"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Pengingat Puasa"
                                    },
                                    "message": {
                                        "type": "string",
                                        "example": "Jangan lupa sahur pukul 04:30"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "reminder"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Notifikasi berhasil dibuat"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/notifications/{id}": {
            "delete": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Menghapus notifikasi berdasarkan ID",
                "operationId": "54c4eb5ec5dc0d1e2916cec2fced1056",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notifikasi berhasil dihapus"
                    },
                    "404": {
                        "description": "Notifikasi tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/report": {
            "get": {
                "tags": [
                    "Report"
                ],
                "summary": "Menampilkan laporan kesehatan pengguna",
                "operationId": "getUserHealthReport",
                "responses": {
                    "200": {
                        "description": "Laporan kesehatan berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "calories_consumed": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 2100
                                        },
                                        "calories_burned": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 500
                                        },
                                        "weight_logs": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "user_id": {
                                                        "type": "integer",
                                                        "example": 123
                                                    },
                                                    "weight": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 70.5
                                                    },
                                                    "recorded_date": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "2025-03-24"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-03-24T14:00:00Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-03-24T14:00:00Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "description": "Enter token in format (Bearer <your-token>)"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}